LayoutTests/fast/canvas/html-in-canvas/texElementImage2D-basic.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>WebGL texElementImage2D — 6-arg basic upload (TB8)</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
  body { margin: 0; }
  #cvs { display: block; }
  #child { width: 32px; height: 32px; background: rgb(255, 0, 0); }
</style>
<body>
<canvas id="cvs" width="32" height="32" layoutsubtree>
    <div id="child"></div>
</canvas>
<script>
function nextPaint() {
    return new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));
}

promise_test(async () => {
    await nextPaint();
    const canvas = document.getElementById("cvs");
    const child = document.getElementById("child");
    const gl = canvas.getContext("webgl");
    assert_not_equals(gl, null, "WebGL context obtained");

    const tex = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_2D, tex);
    gl.texElementImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, child);

    const fbo = gl.createFramebuffer();
    gl.bindFramebuffer(gl.FRAMEBUFFER, fbo);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, tex, 0);
    assert_equals(gl.checkFramebufferStatus(gl.FRAMEBUFFER), gl.FRAMEBUFFER_COMPLETE,
        "FBO complete after texElementImage2D");

    const pixels = new Uint8Array(32 * 32 * 4);
    gl.readPixels(0, 0, 32, 32, gl.RGBA, gl.UNSIGNED_BYTE, pixels);

    // Centre pixel — robust to any aliasing at edges.
    const i = (16 * 32 + 16) * 4;
    assert_equals(pixels[i], 255, "centre red");
    assert_equals(pixels[i + 1], 0, "centre green");
    assert_equals(pixels[i + 2], 0, "centre blue");
    assert_equals(pixels[i + 3], 255, "centre alpha");
}, "texElementImage2D(target, level, internalformat, format, type, source) uploads element pixels at natural size");
</script>
</body>
